refactor: 提取共享 UI 组件，消除 DropZone/DetailField/DetailCard 重复
- 新建 components/shared/DropZone.tsx：统一拖拽上传区域
- 新建 components/shared/DetailField.tsx：详情页字段组件
- 新建 components/shared/DetailCard.tsx：详情页卡片组件
- OcrUploader/FileUploader：移除本地 DropZone，改用共享组件
- ContractDetail/CaseDetail：移除本地 DetailField/DetailCard，改用共享组件
