+++ frontend/src/features/contracts/components/ContractDetail.tsx
@@ -21,6 +21,8 @@ import {
   Sheet, SheetContent, SheetHeader, SheetTitle, SheetDescription,
 } from '@/components/ui/sheet'
 import { PATHS, generatePath } from '@/routes/paths'
+import { DetailField } from '@/components/shared/DetailField'
+import { DetailCard } from '@/components/shared/DetailCard'
 import { formatAmount, formatAmountInt } from '@/lib/format'
 import { downloadBlob } from '@/lib/download'
 
@@ -41,31 +43,6 @@ import {
 export interface ContractDetailProps { contractId: string }
 
 /* ── Shared helpers ── */
-
-function DetailField({ label, value, mono }: { label: string; value: React.ReactNode; mono?: boolean }) {
-  return (
-    <div>
-      <div className="text-muted-foreground mb-0.5 text-xs">{label}</div>
-      <div className={`text-[13px] ${mono ? 'font-mono' : ''}`}>{value || '—'}</div>
-    </div>
-  )
-}
-
-function DetailCard({ title, children, extra }: { title: string; children: React.ReactNode; extra?: React.ReactNode }) {
-  return (
-    <div className="rounded-lg border border-border/60 p-[18px] mb-4 bg-card">
-      {extra ? (
-        <div className="flex items-center justify-between mb-3.5">
-          <h3 className="text-sm font-semibold text-foreground">{title}</h3>
-          {extra}
-        </div>
-      ) : (
-        <h3 className="text-sm font-semibold text-foreground mb-3.5">{title}</h3>
-      )}
-      {children}
-    </div>
-  )
-}
 
 function StatusBadge({ status, label }: { status: string | null; label?: string | null }) {
   if (!status) return <span className="inline-flex items-center px-2 py-0.5 rounded-full text-[11px] font-medium bg-muted text-muted-foreground">未设置</span>
